Fix issue where the flush background task is ended too quick
The background task was ended immediately after it's started, which
potentially would cause the analytic events not being sent upon the app
going to the background.

I'm not entirely sure if this will cause any issue in real apps, but
nonetheless we should make sure the background task is ended at an
appropriate timing.